created a javascript module file (function.js) and add the handleButtonClick function
