+++ index.html
@@ -15,6 +15,8 @@
   <link rel="preconnect" href="https://fonts.googleapis.com">
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link href="https://fonts.googleapis.com/css2?family=Space+Grotesk:wght@300..700&display=swap" rel="stylesheet">
+  
+  <script type="module" src="main.js"></script>
 </head>
 <body>
   <div class="wrapper" id="time-wrapper">
@@ -48,7 +50,6 @@
       <ion-icon name="add-outline"></ion-icon>
     </button>
   </div>
-<script src="main.js"></script>
 </body>
 
 </html>
